admin: split MessageRow
